+++ components/atoms/ButtonOutlined.tsx
@@ -17,7 +17,7 @@ interface Props {
  */
 function ButtonOutlined({ intent = 'default', children, onClick }: Props) {
   const defaultClassName = clsx(
-    'w-full flex items-center pl-6 h-[54px] lg:h-16 border border-Primay-Blue-300 cursor-pointer rounded-2xl lg:text-xl font-semibold text-Primay-Blue-300 hover:bg-[#f3f3f3]'
+    'w-full flex items-center justify-center h-[54px] lg:h-16 border border-Primay-Blue-300 cursor-pointer rounded-2xl lg:text-xl font-semibold text-Primay-Blue-300 hover:bg-[#f3f3f3]'
   );
 
   const activeClassName = clsx({
